Initial perceptron implementation

+++ perceptron/neural_networks/activation_functs.h
@@ -5,7 +5,7 @@
 
 
 namespace ActivationFuncs {
-    float sigmoid(int x)
+    float sigmoid(float x)
     {
         return 1.0 / (1 + pow(M_E, -x));
     }
